Changed MoreInfoModal to a Screen. All the buttons to confirm, reject or cancel a booking have been moved into the MoreInfo screen
